[Refactor] aushSchema, userSchmea 등 재사용 가능한 스키마를 commonSchema로 분리
